# account-service/openapi.yaml
openapi: 3.0.1
info:
  title: "Account Service - TraderX"
  version: "1.0"
  description: "Account management service for TraderX platform. Provides CRUD operations for accounts and account-user mappings. This service is responsible for managing trading accounts and their user associations."
  contact:
    name: "TraderX Team"
    url: "https://github.com/finos/traderX"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0"
paths:
  /accountuser/:
    get:
      tags:
        - account-user-controller
      operationId: getAllAccountUsers
      description: "Gets all user-account mappings"
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountUser"
    put:
      tags:
        - account-user-controller
      operationId: updateAccountUser
      description: "Update user-account mapping"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountUser"
        required: true
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/AccountUser"
        "400":
          description: "Bad Request - Invalid account user data"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "Not Found - Account user mapping not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - account-user-controller
      operationId: createAccountUser
      description: "Creates a user-account mapping"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountUser"
        required: true
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/AccountUser"
        "400":
          description: "Bad Request - Invalid account user data"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "Not Found - Referenced account not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /account/:
    get:
      tags:
        - account-controller
      operationId: getAllAccount
      description: "Retrieves all accounts"
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/AccountList"
        "400":
          description: "Bad Request - Invalid parameters"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags:
        - account-controller
      operationId: updateAccount
      description: "Updates an account"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Account"
        required: true
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          description: "Bad Request - Invalid account data"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "Not Found - Account not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - account-controller
      operationId: createAccount
      description: "Creates an account"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Account"
        required: true
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          description: "Bad Request - Invalid account data"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /accountuser/{id}:
    get:
      tags:
        - account-user-controller
      operationId: getAccountUserById
      description: "Gets an account-user mapping"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/AccountUser"
        "400":
          description: "Bad Request - Invalid account user ID format"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "Not Found - Account user mapping not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /account/{id}:
    get:
      tags:
        - account-controller
      operationId: getAccountById
      description: "Gets an account by ID"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Account"
                
  /account/{id}/trades:
    get:
      tags:
        - account-controller
      operationId: getRecentTrades
      description: "Gets recent trades for an account"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: days
          in: query
          required: false
          description: "Number of days to look back (default: 7)"
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 365
            default: 7
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecentTradesResponse"
        "400":
          description: "Bad Request - Invalid account ID or days parameter"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "Not Found - Account not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "400":
          description: "Bad Request - Invalid account ID format"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "Not Found - Account not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    AccountList:
      type: array
      items:
        $ref: "#/components/schemas/Account"
      example:
        - id: 22214
          displayName: "test account 4"
        - id: 22215
          displayName: "test account 5"
    AccountUser:
      type: object
      required:
        - accountId
        - username
      properties:
        accountId:
          type: integer
          format: int32
          description: "ID of the associated account"
          example: 22214
        username:
          type: string
          description: "Username of the associated user"
          example: "trader01"
          minLength: 1
          maxLength: 50
    Account:
      type: object
      required:
        - displayName
      properties:
        id:
          type: integer
          format: int32
          description: "Unique identifier for the account"
          example: 22215
          readOnly: true
        displayName:
          type: string
          description: "Human-readable name for the account"
          example: "Trading Account 5"
          minLength: 1
          maxLength: 100
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: "Error code identifying the type of error"
          example: "NOT_FOUND"
          enum:
            - "BAD_REQUEST"
            - "NOT_FOUND"
            - "INTERNAL_ERROR"
        message:
          type: string
          description: "Human-readable error message"
          example: "Account not found"
    RecentTradesResponse:
      type: array
      items:
        type: object
        properties:
          id:
            type: string
            description: "Trade identifier"
            example: "ABC-123-XYZ"
          accountId:
            type: integer
            format: int32
            description: "Account identifier"
            example: 22214
          security:
            type: string
            description: "Security symbol"
            example: "ADBE"
          side:
            type: string
            description: "Trade side (Buy/Sell)"
            example: "Buy"
          state:
            type: string
            description: "Trade state"
            example: "New"
          quantity:
            type: integer
            format: int32
            description: "Trade quantity"
            example: 100
          created:
            type: string
            format: date-time
            description: "Trade creation timestamp"
          updated:
            type: string
            format: date-time
            description: "Last update timestamp"

  securitySchemes:
    cors:
      type: apiKey
      name: Origin
      in: header
      description: "CORS enabled. All origins are allowed (*). This service accepts requests from any origin."